Using macro's to avoid duplicated implementations.
